api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
python-jose==3.3.0